:recycle: refactor: 게시글 삭제 변경에 따른 댓글 Service method 변경

+++ src/main/java/com/plane/comment/repository/CommentRepository.java
@@ -13,7 +13,7 @@ import jakarta.validation.Valid;
 
 public interface CommentRepository {
 
-	List<CommentResponse> selectCommentByArticleId(String userId, int articleId);
+	List<CommentResponse> selectCommentByArticleId(String userId, Long articleId);
 
 	int insertComment(String userId, CommentRequest commentRequest);
